- Added support for arm_64 architecture. This is completely transparent to the user, simply upload a arm_64 build of ESXi and it will install it. - Project Monterey (SmartNIC) support added, since we can now boot arm_64. Validated card is NVIDIA BlueField-2. A requirement is to specify mpx device option ( mpx.vmhba64:C0:T0:L0 for BF2 eMMC ). - Added safeguards in both UI and API to not be able to remove Groups, Pools or Images that are currently in use. - Added syslog firewall configuration. Before it would just configure syslog but since firewall is default closed for outbound syslog traffic, nothing would reach the syslog server. - Fixed so that servers using nic other than vmnic0 will now work. both boot.cfg and ks.cfg will now specify to use vnmic of the mac-address defined to boot from as nic to add to vSwitch0. - Added 'Create VMFS' option that will consume the remaining space of the boot drive and create 'datastore1'.

+++ api/groups.go
@@ -210,7 +210,7 @@ func DeleteGroup(c *gin.Context) {
 
 	// Load the item
 	var item models.Group
-	if res := db.DB.First(&item, id); res.Error != nil {
+	if res := db.DB.Preload("Address").First(&item, id); res.Error != nil {
 		if errors.Is(res.Error, gorm.ErrRecordNotFound) {
 			Error(c, http.StatusNotFound, fmt.Errorf("not found")) // 404
 		} else {
@@ -219,11 +219,16 @@ func DeleteGroup(c *gin.Context) {
 		return
 	}
 
-	// Save it
-	if res := db.DB.Delete(&item); res.Error != nil {
-		Error(c, http.StatusInternalServerError, res.Error) // 500
-		return
+	// check if the group is empty, if it's not, deny the delete.
+	if len(item.Address) < 1 {
+		// Delete it
+		if res := db.DB.Delete(&item); res.Error != nil {
+			Error(c, http.StatusInternalServerError, res.Error) // 500
+			return
+		}
+		c.JSON(http.StatusNoContent, gin.H{}) //204
+	} else {
+		c.JSON(http.StatusConflict, "the group is not empty, please delete all hosts first.")
 	}
 
-	c.JSON(http.StatusNoContent, gin.H{}) //204
 }
